Adding Python Binding APIs for fast_sigmoid

+++ extension_cpp/csrc/include/extension_cpp/fast_sigmoid.h
@@ -6,8 +6,8 @@ namespace extension_cpp {
 //swe_cpp_hometest.md: For simplicity, assume `fast_sigmoid` runs on CPU and accepts float32 input only.
 TORCH_API torch::Tensor fast_sigmoid( 
     const torch::Tensor& input,
-    float min_val = -10.0f, 
-    float max_val = 10.0f, 
+    double min_val = -10.0, 
+    double max_val = 10.0, 
     int64_t num_entries = 1000
 );
